+++ CHANGELOG.md
@@ -121,6 +121,7 @@ List of changes for each released npm package version.
       Unreleased changes that have landed in master. Click to see more.
     </summary>
 
+    - [feat] Add ability to collect strings from a custom `FbtCollector` module for collection script
   </details>
 
 - 0.17.2:
@@ -129,7 +130,7 @@ List of changes for each released npm package version.
 
 - 0.17.1:
   - [chore] Additional flow typing in scripts
-  
+
 - 0.17.0:
   - [chore!] `collectFBT` renamed to `collectFbt` (BREAKING CHANGE: updates paths to binary)
   - [minor!] Add ability to write Flow annotations in JS code directly.